+++ client/app/actions/index.js
@@ -19,14 +19,15 @@ import {
   HIDE_NEW_USER_HINTS,
   SHOW_TRASH,
   HIDE_TRASH,
-  TOGGLE_MARK_FOR_KICK
+  TOGGLE_MARK_FOR_KICK,
+  ROOM_STATE_FETCHED
 } from './types';
 import clientSettingsStore from '../store/clientSettingsStore';
 import readDroppedFile from '../services/readDroppedFile';
 import findNextStoryIdToEstimate from '../services/findNextStoryIdToEstimate';
 
 /**
- * store current pathname in our redux store, join or leave room if necessary
+ * Store current pathname in our redux store, join or leave room if necessary
  */
 export const locationChanged = (pathname) => (dispatch, getState, sendCommand) => {
   const state = getState();
@@ -52,11 +53,21 @@ export const locationChanged = (pathname) => (dispatch, getState, sendCommand) =
   });
 };
 
+export const onSocketConnect = () => (dispatch, getState, sendCommand) => {
+  const roomId = getState().roomId;
+
+  if (roomId) {
+    // the socket connected. since we have a roomId in our client-side state, we can assume this is a "re-connect"
+    // make sure we are in sync again with the backend state. send a joinRoom command.
+    joinRoom(roomId)(dispatch, getState, sendCommand);
+  }
+};
+
 /**
  *
  * @param event
  */
-export const eventReceived = (event) => (dispatch) => {
+export const eventReceived = (event) => (dispatch, getState) => {
   const matchingType = EVENT_ACTION_TYPES[event.name];
   if (!matchingType) {
     log.error(`Unknown incoming event type ${event.name}. Will not dispatch a specific action.`);
@@ -78,6 +89,10 @@ export const eventReceived = (event) => (dispatch) => {
 
   if (event.name === 'joinedRoom') {
     history.push('/' + event.roomId);
+  }
+
+  if (matchingType === EVENT_ACTION_TYPES.commandRejected) {
+    tryToRecoverOnRejection(event, dispatch, getState);
   }
 };
 
@@ -347,6 +362,44 @@ export const fetchStatus = () => (dispatch) => {
     dispatch({
       type: STATUS_FETCHED,
       status: response.data
+    });
+  });
+};
+
+/**
+ * If a command failed, the server sends a "commandRejected" event.
+ * From some rejections, we might be able to recover by reloading the room state from the backend.
+ * Obviously there are situations where there is a mismatch between server and client state.
+ */
+const tryToRecoverOnRejection = (event, dispatch, getState) => {
+  if (!event.payload || !event.payload.command) {
+    return;
+  }
+
+  const failedCommandName = event.payload.command.name;
+
+  if (
+    failedCommandName === 'giveStoryEstimate' ||
+    failedCommandName === 'clearStoryEstimate' ||
+    failedCommandName === 'newEstimationRound' ||
+    failedCommandName === 'reveal' ||
+    failedCommandName === 'kick'
+  ) {
+    fetchCurrentRoom(dispatch, getState);
+  }
+};
+
+const fetchCurrentRoom = (dispatch, getState) => {
+  const state = getState();
+
+  if (!state.roomId) {
+    return;
+  }
+
+  axios.get('/api/room/' + state.roomId).then((response) => {
+    dispatch({
+      type: ROOM_STATE_FETCHED,
+      room: response.data
     });
   });
 };