+++ client/app/components/Room/RoomFooter.js
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import appConfig from '../../services/appConfig';
+import {formatDateTime} from '../../services/timeUtil';
+
+import {StyledRoomFooter} from './_styled';
+
+const RoomFooter = () => (
+  <StyledRoomFooter>
+    <div className="version-info">
+      {appConfig.env === 'dev' && (
+        <span>
+          v{appConfig.version}&nbsp;{appConfig.vcsInfo.branch}&nbsp;{appConfig.vcsInfo.hash}&nbsp;
+        </span>
+      )}
+
+      {appConfig.env !== 'dev' && <span>v{appConfig.version}&nbsp;</span>}
+      {formatDateTime(appConfig.buildTime)}
+    </div>
+  </StyledRoomFooter>
+);
+
+export default RoomFooter;