+++ client/app/components/Settings/CardConfigEditorItem.js
@@ -1,0 +1,68 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import {StyledCCTableCell, StyledColorBadge} from './_styled';
+
+/**
+ * one item in the editor. consists of three input fields for label,value,color as well as "move" and "delete" buttons
+ */
+const CardConfigEditorItem = ({item, isLast, isFirst, onChange, onUp, onDown, onDelete}) => {
+  return (
+    <React.Fragment>
+      <StyledCCTableCell style={{width: '25%'}}>
+        <input type="text" defaultValue={item.label} onBlur={onValueChange.bind(this, 'label')} />
+      </StyledCCTableCell>
+      <StyledCCTableCell style={{width: '10%'}}>
+        <input
+          type="number"
+          step="0.5"
+          defaultValue={item.value}
+          onBlur={onValueChange.bind(this, 'value')}
+        />
+      </StyledCCTableCell>
+      <StyledCCTableCell style={{width: '40%'}}>
+        <StyledColorBadge color={item.color} />
+        <input type="text" defaultValue={item.color} onBlur={onValueChange.bind(this, 'color')} />
+      </StyledCCTableCell>
+      <StyledCCTableCell style={{width: '24%'}}>
+        <button
+          type="button"
+          className="pure-button"
+          onClick={onUp}
+          style={{visibility: isFirst ? 'hidden' : 'visible'}}
+        >
+          <i className="icon-up-dir"></i>
+        </button>
+        <button
+          type="button"
+          className="pure-button"
+          onClick={onDown}
+          style={{visibility: isLast ? 'hidden' : 'visible'}}
+        >
+          <i className="icon-down-dir"></i>
+        </button>
+        <button type="button" className="pure-button" onClick={onDelete}>
+          <i className="icon-trash"></i>
+        </button>
+      </StyledCCTableCell>
+    </React.Fragment>
+  );
+
+  function onValueChange(propName, e) {
+    const changedItem = {...item};
+    changedItem[propName] = e.target.value;
+    onChange(changedItem);
+  }
+};
+
+CardConfigEditorItem.propTypes = {
+  item: PropTypes.object,
+  isLast: PropTypes.bool,
+  isFirst: PropTypes.bool,
+  onChange: PropTypes.func,
+  onUp: PropTypes.func,
+  onDown: PropTypes.func,
+  onDelete: PropTypes.func
+};
+
+export default CardConfigEditorItem;