+++ client/app/store/configureStore.js
@@ -3,8 +3,9 @@ import thunkMiddleware from 'redux-thunk';
 
 import rootReducer from '../services/rootReducer';
 import hubFactory from '../services/hub';
-import {locationChanged, eventReceived} from '../actions';
+import {locationChanged, eventReceived, onSocketConnect} from '../actions';
 import history from '../services/getBrowserHistory';
+import appConfig from '../services/appConfig';
 
 /**
  * configures and sets up the redux store.
@@ -13,7 +14,10 @@ import history from '../services/getBrowserHistory';
  */
 export default function configureStore(initialState) {
   const composeEnhancers =
-    (typeof window !== 'undefined' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
+    (typeof window !== 'undefined' &&
+      appConfig.env === 'dev' &&
+      window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+    compose;
 
   let hub;
 
@@ -21,7 +25,10 @@ export default function configureStore(initialState) {
     rootReducer,
     initialState,
     composeEnhancers(
-      applyMiddleware(thunkMiddleware.withExtraArgument((cmd) => hub.sendCommand(cmd)))
+      applyMiddleware(
+        // "sendCommand" will be available in redux action creators as third argument
+        thunkMiddleware.withExtraArgument((cmd) => hub.sendCommand(cmd))
+      )
     )
   );
 
@@ -30,7 +37,12 @@ export default function configureStore(initialState) {
   // All backend events that are received by the hub are dispatched to our redux store
   hub.onEvent((event) => boundActions.eventReceived(event));
 
-  const boundActions = bindActionCreators({locationChanged, eventReceived}, store.dispatch);
+  hub.onConnect(() => boundActions.onSocketConnect());
+
+  const boundActions = bindActionCreators(
+    {locationChanged, eventReceived, onSocketConnect},
+    store.dispatch
+  );
 
   // "sync" url changes to our redux store. if location changes -> store pathname in state
   history.listen(({location}) => boundActions.locationChanged(location.pathname));