+++ server/src/commandHandlers/setCardConfig.js
@@ -1,6 +1,8 @@
 /**
  * A user sets a custom card configuration on the room.
  */
+import defaultCardConfig from '../defaultCardConfig';
+
 const schema = {
   allOf: [
     {
@@ -13,7 +15,7 @@ const schema = {
           properties: {
             cardConfig: {
               type: 'array',
-              minItems: 1,
+              minItems: 0,
               items: {
                 type: 'object',
                 properties: {
@@ -54,12 +56,15 @@ export default setCardConfigCommandHandler;
 
 /**
  * at this point, the command passed validation.
- * cardConfig is an array of objects with "label" "value" and "color" properties.
+ * cardConfig is an array of objects with "label" "value" and "color" properties. (or an empty array).
  * We allow "value" passed as string. try to parse them to number
  *
  * @param {object[]} cc
  */
 function sanitizeCardConfig(cc) {
+  if (!cc || cc.length < 1) {
+    return defaultCardConfig;
+  }
   return cc.map((cardConfigItem) => {
     if (typeof cardConfigItem.value === 'string') {
       cardConfigItem.value = parseFloat(cardConfigItem.value);