+++ server/test/integration/serverIntegrationTest.js
@@ -1,6 +1,7 @@
 import http from 'http';
 import {v4 as uuid} from 'uuid';
 import socketIoClient from 'socket.io-client';
+import poinzSocketClientFactory from './poinzSocketClient';
 
 /**
  * NOTE: for these integration tests, the PoinZ server must be running
@@ -11,7 +12,7 @@ const backendUrl = 'http://localhost:3000';
 describe('websocket endpoint', () => {
   let socket;
 
-  afterAll(() => {
+  afterEach(() => {
     if (socket) {
       socket.disconnect();
     }
@@ -50,112 +51,166 @@ describe('websocket endpoint', () => {
   // we don't want to test all our commands + events here.
   // instead assert one roundtrip -> send valid command -> receive expected events
   // command handling for all commands is tested in unit tests (see test/unit/commands)
-  test('should handle a valid command', (done) => {
-    socket = socketIoClient(backendUrl);
+  test('should handle a valid command', async () => {
+    const client = poinzSocketClientFactory(backendUrl);
 
-    const commandId = uuid();
+    const roomId = uuid();
+    const firstUserId = uuid();
 
-    let eventCount = 0;
+    const receivedEvents = await client.cmdAndWait(client.cmds.joinRoom(roomId, firstUserId), 3);
+    client.disconnect();
 
-    socket.on('event', (msg) => {
-      expect(msg).toBeDefined();
-
-      expect(msg.name).not.toBe('commandRejected');
-
-      expect(msg.correlationId).toEqual(commandId);
-      expect(msg.roomId).toBeDefined();
-
-      if (eventCount === 0) {
-        expect(msg.name).toEqual('roomCreated');
-        expect(msg.userId).toBeDefined();
-      } else if (eventCount === 1) {
-        expect(msg.name).toEqual('joinedRoom');
-        expect(msg.userId).toBeDefined();
-        expect(msg.payload.users).toBeDefined();
-
-        done();
-      }
-
-      eventCount++;
-    });
-
-    socket.on('connect', () =>
-      socket.emit('command', {
-        id: commandId,
-        name: 'joinRoom',
-        roomId: 'my-custom-room_' + uuid(),
-        payload: {}
-      })
-    );
+    expect(receivedEvents[0].name).toBe('roomCreated');
+    expect(receivedEvents[0].userId).toBeDefined();
+    expect(receivedEvents[1].name).toBe('joinedRoom');
+    expect(receivedEvents[1].userId).toBe(firstUserId);
+    expect(receivedEvents[1].payload.users).toBeDefined();
   });
 });
 
 describe('REST endpoint', () => {
-  test('should return backend status', (done) => {
-    httpGetJSON(
-      {
-        host: 'localhost',
-        port: 3000,
-        path: '/api/status',
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      },
-      (err, statusCode, body) => {
-        expect(err).toBeFalsy();
-        expect(statusCode).toBe(200);
-        expect(body.roomCount).toBeDefined();
-        expect(body.rooms.length).toBeDefined();
-        done();
+  test('should return backend status', async () => {
+    const {statusCode, body} = await httpGetJSON({
+      host: 'localhost',
+      port: 3000,
+      path: '/api/status',
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
       }
-    );
+    });
+
+    expect(statusCode).toBe(200);
+    expect(body.roomCount).toBeDefined();
+    expect(body.rooms.length).toBeDefined();
   });
 
-  test('should return 404 if room does not exist', (done) => {
-    httpGetJSON(
-      {
-        host: 'localhost',
-        port: 3000,
-        path: '/api/room/1234',
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      },
-      (err, statusCode, body) => {
-        expect(err).toBeFalsy();
-        expect(statusCode).toBe(404);
-        expect(body.message).toBeDefined();
-        done();
+  test('should export room as json', async () => {
+    // first make sure a room exists
+    const roomId = uuid();
+    const client = poinzSocketClientFactory(backendUrl);
+    await client.cmdAndWait(client.cmds.joinRoom(roomId, uuid()), 3);
+    client.disconnect();
+
+    // export the room
+    const {statusCode, body} = await httpGetJSON({
+      host: 'localhost',
+      port: 3000,
+      path: '/api/export/room/' + roomId,
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
       }
-    );
+    });
+
+    expect(statusCode).toBe(200);
+    expect(body).toMatchObject({
+      roomId,
+      stories: []
+    });
   });
+
+  test('should return 404 if room does not exist (export)', async () => {
+    const {statusCode, body} = await httpGetJSON({
+      host: 'localhost',
+      port: 3000,
+      path: '/api/export/room/1234',
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    expect(statusCode).toBe(404);
+    expect(body.message).toBeDefined();
+  });
+
+  test('should return whole room state', async () => {
+    // first make sure a room exists
+    const roomId = uuid();
+    const userId = uuid();
+    const client = poinzSocketClientFactory(backendUrl);
+    await client.cmdAndWait(client.cmds.joinRoom(roomId, userId), 3);
+    const [storyAdded] = await client.cmdAndWait(
+      client.cmds.addStory(roomId, userId, 'test-story'),
+      2
+    );
+    await client.cmdAndWait(
+      client.cmds.giveEstimate(roomId, userId, storyAdded.payload.storyId, 3),
+      2
+    );
+    client.disconnect();
+
+    // fetch the whole room state
+    const {statusCode, body} = await httpGetJSON({
+      host: 'localhost',
+      port: 3000,
+      path: '/api/room/' + roomId,
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    expect(statusCode).toBe(200);
+    expect(body).toMatchObject({
+      autoReveal: true,
+      id: roomId,
+      selectedStory: storyAdded.payload.storyId,
+      stories: [
+        {
+          consensus: 3,
+          estimations: {
+            [userId]: 3
+          },
+          id: storyAdded.payload.storyId,
+          revealed: true,
+          title: 'test-story'
+        }
+      ],
+      users: [
+        {
+          avatar: 0,
+          id: userId
+        }
+      ]
+    });
+  });
+
+  /**
+   * helper method to send HTTP GET requests to the backend under test.
+   * parsed returned JSON body to object
+   *
+   * @param options
+   */
+  function httpGetJSON(options) {
+    return httpGet(options).then((result) => ({
+      ...result,
+      body: JSON.parse(result.body)
+    }));
+  }
 
   /**
    * helper method to send HTTP GET requests to the backend under test
    * @param options
-   * @param callback
    */
-  function httpGetJSON(options, callback) {
-    const req = http.request(options, (res) => {
-      let output = '';
-      res.setEncoding('utf8');
-      res.on('data', (chunk) => (output += chunk));
-      res.on('end', () => {
-        let parsedBody;
-
-        try {
-          parsedBody = JSON.parse(output);
-        } catch (e) {
-          callback(e);
-        }
-
-        callback(null, res.statusCode, parsedBody);
+  function httpGet(options) {
+    return new Promise((resolve, reject) => {
+      const req = http.request(options, (res) => {
+        let output = '';
+        res.setEncoding('utf8');
+        res.on('data', (chunk) => (output += chunk));
+        res.on('end', () => {
+          try {
+            resolve({statusCode: res.statusCode, body: output});
+          } catch (parseError) {
+            reject(parseError);
+          }
+        });
       });
-    });
 
-    req.on('error', callback);
-    req.end();
+      req.on('error', reject);
+      req.end();
+    });
   }
 });